chore: replace system colors with using app's theme colors, move formatDuration to utils file
