+++ lib/app/features/feed/create_story/views/components/recording_indicator.dart
@@ -2,6 +2,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:ice/app/extensions/extensions.dart';
+import 'package:ice/app/utils/date.dart';
 import 'package:ice/generated/assets.gen.dart';
 
 class RecordingIndicator extends StatelessWidget {
@@ -10,13 +11,6 @@ class RecordingIndicator extends StatelessWidget {
     super.key,
   });
   final Duration recordingDuration;
-
-  String formatDuration(Duration duration) {
-    String twoDigits(int n) => n.toString().padLeft(2, '0');
-    final minutes = twoDigits(duration.inMinutes.remainder(60));
-    final seconds = twoDigits(duration.inSeconds.remainder(60));
-    return '$minutes:$seconds';
-  }
 
   @override
   Widget build(BuildContext context) {
